requirements.txt
numpy==1.26.4
scipy==1.11.4
imutils==0.5.4
opencv-python==4.9.0.80
mediapipe==0.10.11
playsound==1.2.2
Pillow==10.4.0
